+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 ### Fixed
 - Fixed a bug where the plugin was not properly getting all accounts, properties, and views from Google Analytics. ([#10](https://github.com/dukt/analytics/issues/10))
+- Fixed a bug where the Report field was trying to get a data report for new elements. ([#13](https://github.com/dukt/analytics/issues/13))
 
 ## 4.0.5 - 2018-10-26